More Code Clean-up in the Measure Classes
This commit moves some more variables from public to protected in the
Base object, since those variables are not intended to be directly
accessible by other classes like the Manager class.  It's a project
for another day to decide which Base class variables can easily be
made private rather than protected.  Since the option block for a
.MEASURE line is parsed in the Base object, it was easier to simply
make all of the Base object variables accessible in all of its child
classes.  That might not be the best/safest approach though.

This commit also removes some variables that were defined in the
Base object constructor but then either not used, or that were
duplicative of other variables.

Issue #292
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/292