Add ATAN2 support to Xyce/ADMS
The Accellera Verilog-A standard includes a function "atan2" that is a
standard C function.

This commit adds support for this function and its derivatives (and
its use in analog functions) to Xyce/ADMS.

It has not been tested except to examine the code generation of a
"toy" verilog-a model that uses atan2 in a stupid way (and therefore
cannot be run in Xyce in any meaningful way).

Xyce#279
GH#41
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/279
https://github.com/Xyce/Xyce/issues/41