Fix second derivatives of atan2
My last commit regarding atan2 had an error in the emission of second
derivatives (which are almost never needed and are only used when
ddx() is in use *and* the thing ddx computes is used in
contributions (so its derivatives are needed, too)).

This commit fixes those (but they're still not tested).

GH41
Xyce#279
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/279
https://github.com/Xyce/Xyce/issues/41