Consolidate the RFC_ Variable Setting in WhenAt Class
This commits removes the distinction between the RFC_, rise_,
fall_ and cross_ variables in the FindWhen and DerivativeEvaluation
classes and the contRFC_, contRise_, contFall_ and contCross_
variables in the FindWhenCont and DerivativeEvaluationCont
classes.  There actually is none, ever since support for
RFC_ values < -1 was added to the FindWhen and DerivativeEvaluation
classes.  This allows the setting of the RFC_variable (which is
used by the DERIV-WHEN, FIND-WHEN and WHEN measures) to be moved
into the WhenAT class.

Issue #293
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/293