Fixes the GID resolution for dependent variables to handle ground nodes
The Parallel::Directory is a class that the Topology package uses
to determine which voltage nodes are shared in parallel.  Using
the parallel directory requires that the ground node is not included
as a voltage node because the ground node is not shared across
processors.  Each processor has its own ground node, and it's GID is
already assigned -1.  This is a problem because the queries to get
entries from the parallel directory found in the resolveDependentVars
method do not address the ground node separately.  The result is a
runtime error.

This commit modifies the loop over all the devices to resolve GIDs
for any dependent variables to separate out the ground nodes and
set their GIDs to -1.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/296