Update extractMIDeviceData comments
In commit 643cb85 (related to issue Xyce#6), the parsing of a YMIN or
YMIL device line was updated to include an initial condition as part
of the inductor list data.  This change in function was not reflected
in the comments of the function.

This has no impact at all to most users, but might confuse developers
who are trying to see how the YMIN/YMIL device actually works by
hand-rolling a YMIL line meant to duplicate what the parser is doing
under the hood.