Change some for loops to auto/ranged C++11 style
Xyce requires a fully-compliant C++11 compiler since release
7.3 (actually, since sometime after 7.2).  As such, using C++11
features that do not exist with C++03 compilers is now fair game.

One such feature is the "auto" declaration that lets the compiler
choose the correct declaration for a variable based on how it is used.

The other is ranged for loops that allow one to loop over STL
containers without explicit creation and manipulation of iterators.

This commit changes two functions in N_IO_OutputMgr to make use of
these language features.  Instead of:

   for (OutputParameterMap::const_iterator it1 = outputParameterMap_.begin()
     , end1 = outputParameterMap_.end(); it1 != end1; ++it1)
   {
    const OutputParameterMap::mapped_type &parameter_vector = (*it1).second;
    ...

we can do:

  for (const auto &outputParameterPair : outputParameterMap_)
   {
    const auto &parameter_vector = outputParameterPair.second;

The two functions I modified have a lot of loops like this, and using
the new C++11 syntax makes the code a lot more readable and
maintainable.

Xyce#298
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/298