First fix for coupled inductor outputs
Issue #297
"Problems using coupled inductors (from mutual inductors) on .PRINT and
.SENS lines, amongst other problems"

This issue has to do with accessing parameter L values for coupled
inductors.   In Xyce and other SPICE simulators, mutual inductors are
specified as multiple devices in the netlist:

L1 1 0 1.0
L2 2 0 2.0
K1 L1 L2 0.5

Internally,  Xyce handles mutual inductors by bundling all the
coupled inductors into a single device.  This is convenient in some
regards, particularly parallel.  But, it has the effect of making the
original L devices no longer exist as separate devices, and this makes
it so Xyce's normal parameter handling functions cannot find them.

Fixing this is mostly a book-keeping issue, and requires creating
hackish exceptions in some of the parameter-related functions.  This
commit fixes several use cases, but not all of them.

This commit fixes this issue for outputs such as the .PRINT line.

It also fixes the issue for UQ analyses, so that coupled inductors can
be set via .DC/.STEP/.SAMPLING/.EMBEDDEDSAMPLING/.PCE.

It partially, but not completely, fixes this issue for sensitivity
analysis.  Sensitivity analysis will work on parameters like this, but
only when using the oldest brute force finite difference derivative when
computing the sensitivity residual.

In general, derivatives needed by the residual come from devices, and
they can be computed 3 ways:
1. analytically, if the device has been instrumented to do it that way
2. numerically on the device level (this works fine for std parameters)
3. numerical on the vector level (this is the original, slow method)

The coupled inductor parameters are *not* set up for either 1 or 2.  That
leaves option 3, which is not as nice.  One reason it is not as nice is
that it doesn't work with the sparse version of transient adjoint
sensitivities.    I am hopeful that option 2 can be set up relatively
quickly, but that might not be the case.  I'll have to dig more.