Fix for coupled inductor adjoint sensitivities
Issue #297
"Problems using coupled inductors (from mutual inductors) on .PRINT and
.SENS lines, amongst other problems"

As noted in my previous commit, this work concerns various outputs and
post-processing calculations (such as .SENS) for coupled inductors.
Mutual inductors are specified as multiple devices in a netlist:

L1 1 0 1.0
L2 2 0 2.0
K1 L1 L2 0.5

But are treated as a single monolithic device inside the Xyce source
code.  As a result, the L devices get absorbed into a K device, and as a
result the normal parameter handling machinery is unable to find them.
This broke various output capabilities for coupled inductors.

My previous commit fixed this issue for outputs like .PRINT, but only
partially fixed the issue for sensitivity calculations.  This commit
implements a more comprehensive fix for sensitivities.

In the previous commit, I fixed the device manager so that functions
such as DeviceMgr::setParam and DeviceMgr::getParam would work with
coupled inductors.  This is sufficient to support the least efficient
method for obtaining finite difference residual derivatives. This is
the method invoked by the "forcefd=1" parameter.  But it was not
sufficient to support the more efficient "device level finite
difference", invoked by the "forcedevicefd=1" parameter.

The forcefd=1 method causes the entire residual vector to be re-loaded
and differenced.   In contrast, the forcedevicefd=1 method only
differentiates the bits of the residual written to by the device.  So,
the second method is more compatible with a sparse storage structure,
and is more efficient.  It is also the default finite difference method,
and the only finite difference method compatible with transient
adjoints (which must use sparse storage).

With this commmit the forcedevicefd=1 method now works on coupled
inductors.

The only (possibly) remaining thing to be done on this (maybe) would be
to implement analytical derivatives for the L-params in mutual inductor
devices.  That's probably feasible for linear mutual inductors (not
clear about nonlinear mutual inductors).  But it might not be worth the
effort.