Fix unsafe pointer use in numerical .sens residual
Issue #297
"Problems using coupled inductors (from mutual inductors) on .PRINT and
.SENS lines, amongst other problems"

The code I added to support device-level numerical sensitivities for
coupled inductors had a pointer usage which should have checked first
for a null pointer.  This would not break the code in serial, but would
break it in parallel, since the device level finite difference
calculation will just be local to one processor.

My changes from earlier today broke around 15 sensitivity tests, and
they are all tests that use the device-level finite differencing.  I
have verified that this mistake (null pointer) was the reason for those
failures.  This commit fixes the issue and adds an if-statement to the
pointer use in the DeviceMgr::getNumericalSensitivities function.