Another fix for mutual inductor op builder
Issue #297
"Problems using coupled inductors (from mutual inductors) on .PRINT and
.SENS lines, amongst other problems"

The makeOps structure, which is based in the N_UTL_Op.C file, makeOps
function is very fragile.  For every string on the .PRINT line, it loops
over all the Op builders until it finds one that succeeds, and then
stops.  If one *inappropriately* succeeds, then it stops at the wrong
time.  So, it is very important for Op builder classes to be strict, and
not allocate any ops that aren't quite correct.  I've been bitten by
this before.

I made a mistake of this nature in setting up the
MutualInductorInstancesOpBuilder::makeOp function.

The makeOp function was allocating an Op even if the string was an
expression that happened to contain a coupled inductor.  It should only
have been allocating an op for a pure (non-expression) coupled inductor
variable.  In other words if you have:

.print tran X1:L1:L  {X1:L1:L}  {3.0*X1:L1:L}

then all 3 of the outputs were getting allocated as
MutualInductorInstancesOps.  Only the first one should have been.  The
other 2 should have been allocated as expression Ops, and then they
should have been allocating MutualInductorInstancesOps in a later stage,
to support that expression.

One result of this mistake was that for the above .PRINT line, the final
field ({3.0*X1:L1:L}) would output X1:L1:L, which was wrong.  Another
result was that the Ops were inconsistent across processors, which
caused a fatal error in parallel.

This commit fixes the error.  It modifies the
MutualInductorInstancesOpBuilder::makeOp so that it first checks if the
tag is an expression, and if it is, it does not attempt to find the
coupled inductor.