Move More Common Code into the WhenAT Class
This commit changes the functions updateCalculationResult() and
updateCalculationInstant() from pure virtual in the WhenAT class
to being non-virtual protected functions in that WhenAT class.
The code for those functions was identical in the child class,
since negative RFC_ values were supported for the FindWhen and
DerivativeEvaluation classes.

This commit also does some other clean-up of the refactoring.
Suggestions for additional refactoring will be discussed in
Issue #278.

Issue #293 and #278
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/293
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/278