Fix some code to conform to Xyce style
Fixed a couple of "egyptian" brace issues, which are not the Xyce
style:

   if (foo) {
     bar
   }

and morphed them into the Xyce style:
  if (foo)
  {
    bar
  }

Changed a handful of comments referring to something being done in
Python, since the comment really means something is being done in the
external device code (whatever language it is in).

Removed some completely unnecessary braces.

Removed a few whitespace errors.