Cleans up floating node removal and supernoding capabilities
Recent extracted netlists from large ASIC designs have included
a non-negligable number of floating devices that are not attached
to the rest of the circuit.  These devices can, optionally, be
removed from the circuit.  Determination of these devices needs to
be done carefully in parallel.  This commit provides a new cleaned
up version of that capability.

It was also noticed in fixing another bug in the parallel topology
setup that unexpected nodes will result in runtime errors during the
shared voltage node assignment.  While a previous bug was focused on
ground nodes, this is also true for voltage nodes being collapsed
through supernoding.  This commit includes a fix for using aliases
from supernoding in the shared voltage node assignment.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/250