Add known defect for ADMS modulus operator
I have known for some time that Xyce/ADMS does not generate
LRM-compliant code for the modulus operator.  Instead, it just emits
the exact C++ expression using the same opterator (this is wrong).  In
cases where the Verilog-A uses arguments that are not both integer
expressions, the result is code that does not compile.  In cases where
both are integers, it's fine.

This operator is not used in most models, but the HiSIM uses it for
the trivial purpose of extracting the major and minor version number
of the model out of a "real" version number.  There is an open issue
in gitlab for it, but it is low priority.

This came up in an email discussion of the new expression library's
modulus operator, which differs from the old expression library's in
the same way.  It occurred to me that there had been no documentation
of Xyce/ADMS's defect in this regard.

Xyce#41
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/41