Add support for Bison 3.8 to autoconf build
The autoconf build is written to accept only enumerated versions of
bison, and a version of bison is added to the list only when we have
verified that it is backward compatible with the oldest version we
MUST support (which is determined by what ancient version is provided
by the current Sandia version of RHEL).

An open source user has found that Arch Linux has already leapt to
version 3.8 of Bison, which was released in September.  He has tested
it with Xyce and found it to work.  I am therefore adding it to our
release 7.4 lest that version suddenly start appearing on other
platforms and making our release obsolete before it's even out the
door.

https://github.com/Xyce/Xyce/issues/44
GH-44