Fix copyright headers
A few files had outdated copyright ranges, and a few other files had
an obsolete copyright block format (the current format for open source
files is correctly found in the now-misnamed
"src/headers/CVS-C++-File-Header-OS" and for ECI code in
"src/headers/CVS-C++-File-Header-ECI"

As near as I can tell these are the only files that needed updating
for the 7.4 release, and they should have been caught for 7.3.

In some cases, the difference in formatting could be responsible for
them not being updated, because sometimes we do the updates using a
regular expression replacement, and files with incorrectly formatted
headers get missed.  In the case of the wildcard support stuff, it
seems that the format is fine but somehow they got overlook,ed anyway.

Xyce#312
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/312