Remove the binaryModOp tests
Issue #310
"Fix modulus operator in new expression library to properly handle real
number inputs"

As of this writing, this is a "closed" issue.   But in my work on the
modulus operator, I removed the binaryModOp class, as fixing it would
have made it redundant with the fmodOp class.  However, I forgot to also
remove the AST unit tests for binaryModOp.  That rendered these 2
programs (astUniTest and astCopyConstUnitTest) un-compilable.

These are not the most important unit test programs for the expression
library, which is the main reason I forgot to update them.  I mostly run
the parserUnitTest and complexParserUnitTest programs, and those were
updated for the work on this bug.

This commit fixes the AST unit test programs by removing all tests for
binaryModOp.