Add note about model binning not working w/UQ
Issue #316 is now added to the known defects table.
"Model binning and UQ analyses on L,W parameters don't play well
together"

This issue was noticed recently by a user.  As model binning is
relatively new, this particular use case doesn't work yet.  Model
binning is done during the parsing phase and any models that don't
appear to be needed during that phase are not stored.  As a result,
if L,W parameters are updated later, the device package doesn't have
enough information to properly re-bin the models.