Fix SpiceSeparatedFieldTool to avoid unnecessary replacements of ground synonyms
The SpiceSeparatedFieldTool breaks down every line from a netlist
into a vector of strings through the getLine method.  This method is
passed a boolean that replaces any ground synonym ("GND","GND!","GROUND")
with "0", which is the default identifier for the ground node in Xyce.
The boolean passed to the getLine method in the SpiceSeparatedFieldTool
object, replgndvar, results in any ground node synonym being replaced
even if it doesn't make sense to do so.  This includes .SUBCKT statements,
as well as .INCLUDE/.INC/.INCL/.LIB/.ENDL statements.

This commit adds a trap for any directives for which it doesn't make sense
to replace a ground synonym and then sets the replgndvar to false to avoid
replacements in the rest of the line.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/318