Improve Error Checking for Invalid .MEASURE Lines
This commit improves the error checking for invalid DERIV-AT,
DERIV-WHEN, FIND-AT, FIND-WHEN and WHEN measure lines.  It also
removes some duplicate calls to checkMeasureLine() in the
constructors for the IO::Measure::Extrema, IO::Measure::Stats
and IO::Measure::TranStats classes.  Those calls (and the setting
of the typeSupported_ flag) were redundant with the code in
the constructors for the child classes.

Issue #319
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/319