Fixes NaN check in DampedNewton convergence test
The current NaN check in the convergence test uses the inf-norm,
which will not detect a NaN in the residual vector.  That is because
an inf-norm computation will detect the largest value and a NaN is
neither greater than, less than, or equal to another value.  Thus,
it will not be chosen as the resulting value of the inf-norm.

Instead a 2-norm should be used, because arithmetic operations with
a NaN result in a NaN.  Thus, if any value in the residual vector is
a NaN, the resulting value will be a NaN.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/321