Fix for .PREPROCESS REPLACEGROUND w/expr library
Issue #320
"Fix handling of ground nodes in expression library to be consistent
with .PREPROCESS REPLACEGROUND"

I worked on this issue a while ago, but didn't finish it.  Then I kind
of forgot about it until recently.

Prior to the changes in this commit, the expression library assumed that
the string "GND" was synonymous with node 0.  It did this
unconditionally.

However, in Xyce's current design, this should only happen if the
.PREPROCESS REPLACEGROUND command is invoked in the netlist.  Also, when
that command is invoked, it is supposed to support strings GND! and
GROUND as well as GND.

This commit accomplishes both things.  In order to do this, I had to add
the preprocess boolean array to the static objects of the
N_UTL_HspiceBools.h file, add operations on that static object in the
netlist parser, and then add awareness of it in the expression parser
*yxx file.

I've also created several unit tests for this feature, and they are
present in this commit.