Fix reference and users guides references
Another little annoying thing that often gets forgotten until after
R&A:  The UG and RG have references to each other, and these actually
need to be updated on every release.  IN the RG it's done in a macros
file, but in the UG it's right in the main file.

Xyce#312
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/312