Enforce AT vs. TD Precdence for TrigTarg Measures
This commit enforces the precedence of the AT vs. TD qualifiers
for TRIG-TARG measures that use the TrigTarg classes.  The rules
are:

First, if both AT and TD are given in both the TRIG and TARG clauses
then both TD's are ignored.  An example is:

  .MEASURE TRAN AtTD1 TRIG AT=2ms TD=5ms TARG AT=8ms TD=9ms

However, if only the TRIG TD and AT qualifiers are given then that TD
value is still ignored for the TRIG clause.  However, that
TD value is then used for the TARG clause.  This may be confusing
but it does match HSPICE.

  .MEASURE TRAN AtTD2 TRIG AT=1ms TD=2ms TARG V(1)=0.3 CROSS=1

Issues #289 and #302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302