Remove Measure Window Info from stdout for TRIG-TARG
The newer (more HSPICE compliant version) of the TRIG-TARG measure
allows different AT and TD values in the TRIG and TARG clauses.
This complicates the generation of the "measure window bounds" for
display in the stdout.  I had two options: 1) remove that info for
the newer version of TRIG-TARG; or 2) print out separate window
info for the TRIG and TARG clauses.  I went with Option 1, since it
was simpler and matches what HSPICE provides in stdout for the user.

Issues #289 and #302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302