Prep development branch for next release
Configure and cmake set up to show "development" code instead of
release.  For Autoconf build this means that the version string
printed by "Xyce -v" will be the DEVELOPMENT-<date>-<describe> where
date is the date/time of the actual build and describe is the
repository state corresponding to the code being built.  At the
moment, the code that constructs the version string is broken for the
CMake build and <date> is the date/time at which *cmake* was run (not
make itself) and describe is the repository state corresponding to the
same run of cmake, not necessarily the code actually built into the
binary.

I have added a boilerplate reference to Xyce 7.5 user and reference
guides to circuit.bib.  I've changed the guides to say they're 7.5,
and removed their SAND numbers.