More Improvements for TRIG-TARG Measures
This commit improves the handling of failed TRIG-TARG measures, that
use the new TrigTarg classes, for all measure modes (AC, AC_CONT,
DC, DC_CONT, NOISE, NOISE_CONT, TRAN and TRAN_CONT measure modes.
Most of the changes revolved around handling invalid AT or TD values.
Note that Xyce will report a failed window if the AT value (for either
the TRIG or TARG clause) is outside of the simulation window.  This
behavior may differ from HSPICE in some cases.

The second change was in the handling of the TD qualifier for DC
and DC_CONT measure modes.  The TD qualifier is now "independent" of
the sweep direction (ascending or descending).  This matches HSPICE,
but it is different from how FROM, TD and TO are handled for DERIV-WHEN,
FIND-WHEN and WHEN measures.

Finally, the outputters, for the TrigTargCont classes, were also
improved to print out information about whether the TRIG clause, the
TARG clause or both failed.  As an example for TRAN_CONT mode, if the
TRIG fails but the TARG succeeds then the TRIG will be reported
as "not found" and the first TARG value will be reported.  This
is slightly different than HSPICE for continuous measures.  A
failed continuous measure in HSPICE will simply produce no output.

Issue #289 and #302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302