Update Ref. Guide and 7.5 Rel. Notes for Issue 220
This commit adds gitlab-ex issue 220 to the Fixed Defects table in
the Xyce 7.5 Release Notes.  It also updates the "HSPICE
Compatibility" subsection of the .MEASURE section of the Reference
Guide.  Issue 220 was about the TRIG-TARG measure being able to
return a negative value if the TARG condition occurred before the
TRIG condition.

Issue #220
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/220