Support for SPICE-style command, .option scale
Issue #237.
"add parser support for ".option scale" so that ".options parser scale"
doesn't have to be translated via XDM"

Xyce has nominal support for SCALE, where transistor lengths and widths
are automatically scaled by a netlist-given scale factor.  However, my
implementation used a Xyce-style command: ".options parser scale=#".

This was fine, but was not the same as the SPICE style command used by
nearly every other simulator, which is ".option scale=#".  Converting
this is something that the XDM tool can do.    However, not everyone is
using XDM, particuarly the people who are using Xyce with the Sky130
PDK.  Sky130 is in ngspice format, which is very close to Xyce.

This issue is one of the only remaining compatibility issues for Xyce
with Sky130, so it seems worthwhile to support it directly, rather than
rely on conversion scripts and/or XDM.

This commit modifies the following.

In the extractOptionsData function in the N_IO_PkgOptionsMgr.C file,
there was already code to convert ".options tnom/temp=number" to ".options
device tnom/temp=number".  I followed the pattern of that code to now
also convert ".options scale=number" to ".options parser scale=number".

I also added a new call to the function addCommandParser in the
PkgOptionsMgr constructor.   So now, in addition to this:

addCommandParser(".OPTIONS", extractOptionsData);

there is also this:

addCommandParser(".OPTION", extractOptionsData);

This means now that anytime ".OPTION" appears in the netlist, the
function Xyce::IO::extractOptionsData will be called on it, to process
the line.  And it will be treated the same as ".OPTIONS".  So, that
implicitly adds support for things like ".OPTION DEVICE", synonymous
with ".OPTIONS DEVICE".  Although that isn't the point of this issue, I
think that is OK.

The main potential problem was that before this commit, any ".OPTION"
(singular) command in the netlist would be treated as an unrecognized
command, and Xyce would mostly ignore it, while outputting a warning.
Now, any ".OPTION" command, including ones we don't suppport (which
is 99% of SPICE commands), will get extracted in this function.
The old error trap was a fatal error, rather than a warning, so
I've changed it to be a warning.