Change Xyce Behavior for Failed Measures
This commit changes what Xyce prints to the .MEASURE output files
for failed measures.  The default is now to print FAILED in those
files, which corresponds to a default setting of .OPTIONS MEASURE
MEASFAIL=1.  .OPTIONS MEASURE MEASFAIL=0 will print 0 to those
output files for a failed measure.  If .OPTIONS MEASURE MEASFAIL=0
is used then the DEFAULT_VAL qualifier on individual .MEASURE
or the .OPTIONS MEASURE DEFAULT_VAL=<value> will be honored, with
.OPTIONS MEASURE DEFAULT_VAL taking precedence over the DEFAULT_VAL
qualifier.

Issue #304
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/304