modify IO to use last netlist instance of .param
Issue #202.

"When multiple instances of the same .param appear in the netlist, Xyce
uses the first value, while Hspice and ngspice use the last value"

It is common in PDKs for .param to be set more than once.   Typically, a
.param will be set inside of a .lib  (for its default value) and then in
the main circuit it may also be set, to possibly override the default.

Most circuit simulators, including Xyce, allow multiple instances
of the same .param to appear in a netlist.

However, Xyce has handled this use case differently.  Xyce uses the
first instance of a .param it encounters in the netlist, and then
ignores any others it might find.

Hspice,ngpsice and Spectre do the opposite.  They use the *last*
instance of a .param that they find in the netlist.

Xyce has done this for a pretty long time, at least the last several
code releases.  In the parser, Xyce stores all .PARAM in a
std::unordered_set.   All the sorting, find commands, insertions,
etc are based on the name of the parameter.

For std::unordered_set, if you attempt to insert something, but
it matches something that is already in the container, then
the insertion simply doesn't happen.   So, for that reason,
Xyce has used the first version that it finds in the netlist.

In this commit, I modify the code that handles this. Now, Xyce
checks if that parameter is there already.

If the parameter already exists, in my hack I erase the existing parameter
and then insert the new one.  (The members of unordered sets are immutable
(i.e., const) so I can't modify an existing member)

This does not break any existing serial tests.  And for the simple
example test cases I have put together, Xyce now matches ngspice and
Hspice.

In addition to changing Xyce's default behavior, it also gives the
user several options for handling default  parameters, including the old
behavior.  These options are specified on the command line.  I initially
tried to make it so they could be specified via .OPTIONS PARSER.  But, at the
moment, this cannot be made to work, because .OPTIONS and .PARAM are
both processed on the same netlist pass, and there is no way to ensure
that the the .OPTIONS will be set prior to a .PARAM being parsed.

So, for now, this is a command line only option.  The new command line
is:

 Xyce -redefined_params <option>

where option = IGNORE, ERROR, WARNING, USEFIRST or USEFIRSTWARN.

IGNORE is the default behavior, and means that the last instance
of a .PARAM in the netlist will be used.

ERROR means that duplicated params will cause a fatal error.

WARNING is the same as IGNORE, except that each duplicate parameter will
issue a warning.

USEFIRST is included for backward compatibility with recent Xyce
releases, which instead of using the last instance of a parameter in the
netlist, will instead use the first.

USEFIRSTWARN is the same as USEFIRST, except that it issues a warning
for each duplicated .param.

All of the above applies to .global_params as well as .params.