Implements variable masking for DampedNewton nonlinear solver
Devices can indicate internal solution variables that need to
be masked by the simulation engine, often due to scaling.
These device masks are used by the time integrators in Xyce
to "ignore" variables in the determination of convergence.
However, the device masks are not used by the nonlinear solver.
In some cases the nonlinear solver is failing, even if the masked
solution would have been acceptable to the time integrator,
and the time integrator will treat this as a failure
and will use step or order reduction.  This can result in a
slower simulation, where the use of device masks in the nonlinear
solver may enable the time integrator to determine convergence.

This commit includes an implementation of the variable masking
for the DampedNewton nonlinear solver.  The NOX nonlinear solver
technically performs masking, but the implementation is not complete.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329