fix the inefficient initialization for multi-tone HB techniques
The initialization of multi-tone HB takes the majority of the simulation
time, especially if the number of frequencies is large. This is true
for both transient assisted and non transient assisted initial guess
calculation. For the multi-tone HB tests, the initialization took more than 90% of
the simulation if 100 harmonics is used for one tone.

The inefficient initialization is due to the generation of time points
in the APFT algorithm. The APFT algorithm selects the time points which lead
to good condition number of the APFT matrix. This algorithm can be
expensive for large number of frequency points. The transient assited HB
uses transient simulation to calculate initial guess. Because frequency
mapping based HB does not related time and frequency domain information, it
uses APFT matrix to help calculate initial guess. The APFT time point calcuation
is used in both APFT and frequency mapping based multi-tone HB methods.
This leads to very slow initialization. However, for non tahb initial guess
calculation, the frequency mapping based HB does not need the APFT
matrix. This is fixed in this commit. It lead to large speedup in the
frequency mapping based multi-tone HB simulation.

Xyce#326