Update Xyce Guides for TRIG-TARG Measure Improvemenents
This commit contains the complete documentation for issue 302
("Implement AC_CONT, DC_CONT, NOISE_CONT and TRAN_CONT modes for
TRIG-TARG measures").  It is partial documentation for issues
289 ("Add Support for TRIG-TARG measure for AC, DC and NOISE
measure modes") and 303 ("Support both legacy (Xyce 7.4 and before)
and HSPICE compatible TRIG-TARG measures ").  The documentation
for the latter two is complete in the 7.5 Release Notes, the
Users' Guide, and the .OPTIONS section of the Reference Guide.
I still need to update the .MEASURE section of the Reference
Guide for issues 289 and 303.

Issues #289, #302 and #303
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/303