Remove compiler warnings in select source files
These changes are primarily clean up (remove warnings when compiled
with -Wall) following the development work for another issue (#292).
The scope of the clean up work was established in the initial commits
for this issue (02521e8c, 2bc13caa, aa6e5080). This commit is intended
to address the remaining warnings from files impacted in these
previous commits.

There is still a warning about an unused variable that is pulled in
from a nested #include that will require a larger structural change.

Issue #294
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/294