Make BJT excess phase to use store, not state vec
Issue #337
"Switch the BJT excess phase to use the store vector rather than the
state vector"

This came up in the context of another bug.

The excess phase model in the BJT computes a second order time
derivative.  It *can* do this via the DAE formulation, but this causes
there to be 2 extra equations per device and this is not desirable.
This is an option in the model, but not what is usually done.

So, instead of doing this, the device does the calculation completely
inside the device.   It has used the state vector for bookkeeping.

But generally, the state vector is supposed to be used for stuff that is
differentiated.  Alternatively, the store vector is used for stuff that
is simply tracked.  ie, pure bookkeeping should be done with "store"
rather than "state".

If the excess phase model had relied upon the state vector's time
derivatives, then its use would have made sense.  But that isn't how the
model is set up.  And, for the purposes of being compatible with the
original spice3 version, it should stay that way.

This commit updates the BJT model to use the store vector in the excess
phase model.  Doing this required that a longer history of store vectors
be tracked, and this changed the function argument list for a lot of
functions such as "updateState" and "loadDAEVectors".  These changes
propagated thru all the loaders up the to data store class.  This wasn't
hard to do but it required a lot of typing.

In serial, everything passed, so I'm pushing it.