Address max-time-step issue
When simulink tries to set the max time step in a Xyce simulation
it can cause Xyce to hang.  For now I'm not having the simulink
layer attempt to set a max time step and relying on Xyce to pick one.
That appears to always work, but I'll need to dig further into
Xyce to figure out what goes wrong when simulink does try to set a
max time step.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104