Small fix to computeBreakPoint function
Issue #340
"Bug in breakpointing, when an expression combines a ternary and a
table"

While working on a different issue, I discovered a pathological case in
which the "computeBreakPoint" function would return whatever the current
time is.  The specific case where this happened was one in which I was
combining a ternary operator and a (time,value) table.  But, in theory,
it could have happened under other circumstances involving a ternary.

The ternary, essentially said:

(tableValue > 0.0)?(tableValue):(0.5)

This was bad b/c the ternary comparison and the second option didn't
match, which meant this would cause discontinuities.  But, it was also
bad because the specific table had value=0.0 for a lot of time points.A

So, that meant that if you were at one of those time points, then
tableValue==0.0 and the derivative (dfdt) was always zero.   Inside the
computeBreakPoint function's Newton loop, this meant that the Newton
convergence criteria was satisfied before the loop even started, and the
current time would be the breakpoint.

In practice, this pathological case probably won't happen much.  But to
prevent it in the future, I've modified the function to ensure that if
dfdt==0.0, it will not attempt to compute a breakpoint.