Getting size of time-voltage vec doesn't clear it
In the first implementation of the getTimeVoltagePairsSz() function
I didn't realized that the ADC device automatically clears the time-
voltage history after a calling program has requested it.  Thus
checking the size of the history also cleared the history.  Now,
the size checking doesn't clear history.  But getting the actual
history vector does clear it just to maintain the old (and odd)
behavior.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/338